backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
google-generativeai==0.3.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
drf-spectacular==0.26.5
